\name{capwire-package}
\alias{capwire-package}
\alias{capwire}
\docType{package}
\title{
Estimate Population Size from Non-Invasive Sampling
}
\description{
capwire uses capture-recapture models which allow for sampling individuals multiple times per session to estimate population size from non-invasive samples. These models were originally implemented in the Java program Capwire (Miller et al. 2005)

The models all assume that individuals can be identified with certainty (usually with, but not limited to) genetic data.

The package also includes functions to compare models, get confidence intervals using a parametric bootstrap and simulate data under a variety of conditions.
}
\details{
\tabular{ll}{
Package: \tab capwire\cr
Type: \tab Package\cr
Version: \tab 1.1.5\cr
Date: \tab 2013-08-02\cr
License: \tab License: GPL (>= 2)\cr
}
}

\author{
Matthew W. Pennell and Craig R. Miller

Maintainer: Matthew W. Pennell <mwpennell@gmail.com>
}

\references{
Miller C. R., P. Joyce and L.P. Waits. 2005. A new method for estimating the size of small populations from genetic mark-recapture data. Molecular Ecology 14:1991-2005.

Pennell M.W., C.R. Stansbury, L.P. Waits and C.R. Miller. submitted. capwire: A R Package for Estimating Population Census Size from Non-Invasive Genetic Sampling 
}
