% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{capwire}
\alias{capwire}
\alias{capwire-capwire}
\alias{capwire-package}
\title{Capwire: estimate population census size from non-invasive genetic sampling}
\description{
This R package is designed to fit the sampling models developed by Miller et al. 2005 (Molecular Ecology). The two main model fitting functions are \code{\link{fitTirm}} and \code{\link{fitEcm}}. The package also includes tools to estimate confidence intervals by parameteric bootstrapping (\code{\link{bootstrapCapwire}}), perform model selection (\code{\link{lrtCapwire}}) and simulate data under a variety of conditions (\code{\link{simTirm}}, \code{\link{simEcm}}, and \code{\link{simCapture}}).
}
\references{
Miller C. R., P. Joyce and L.P. Waits. 2005. A new method for estimating the size of small populations from genetic mark-recapture data. Molecular Ecology 14:1991-2005.

Pennell, M. W., C. R. Stansbury, L. P. Waits and C. R. Miller. 2013. Capwire: a R package for estimating population census size from non-invasive genetic sampling. Molecular Ecology Resources 13:154-157.
}

