% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawCapRatesBeta}
\alias{drawCapRatesBeta}
\title{Create a exponential distribution of capture rates for simulation}
\usage{
drawCapRatesBeta(shape1, shape2)
}
\arguments{
\item{shape1}{first shape parameter of beta distribution}

\item{shape2}{second shape parameter of beta distribution}
}
\value{
a function which takes an argument \code{n} that can be then used to simulate capture rates with \code{\link{simCapture}}
}
\description{
Can be used in conjunction with \code{\link{simCapture}}
}
\examples{
## Specify a beta distribution
bd <- drawCapRatesBeta(1, 0.5)
simCapture(n=20, s=100, bd)
}
\seealso{
\code{\link{simCapture}} \code{\link{rbeta}}
}

