% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawCapRatesGamma}
\alias{drawCapRatesGamma}
\title{Create a gamma distribution of capture rates for simulation}
\usage{
drawCapRatesGamma(shape, rate)
}
\arguments{
\item{shape}{gamma shape parameter}

\item{rate}{gamma rate parameter}
}
\value{
a function which takes an argument \code{n} that can be then used to simulate capture rates with \code{\link{simCapture}}
}
\description{
Can be used in conjunction with \code{\link{simCapture}}
}
\examples{
## Specify a gamma distribution
gd <- drawCapRatesGamma(1,0.5)
simCapture(n=20, s=100, gd)
}
\seealso{
\code{\link{simCapture}} \code{\link{rgamma}}
}

