% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simEcm}
\alias{simEcm}
\title{Simulate data under Equal Capture Model (TIRM)}
\usage{
simEcm(n, s)
}
\arguments{
\item{n}{Number of individuals in the population}

\item{s}{Total number of samples collected}
}
\value{
A two-column matrix with the first column specifiying the capture class (i.e. individuals caught i times) and the second column specifying the number of individuals in each class.

The matrix can be used as the data argument for any of the model-fitting functions implemented in capwire
}
\description{
Simulates capture count data under the assumptions of the Equal Capture Model (ECM) where all individuals in the population are assumed to be equally likely to be captured.
}
\examples{
## Simualte data under Equal Capture Model

data <- simEcm(n=25, s=100)

## Fit the Equal Capture Model

ecm <- fitEcm(data=data, max.pop=200)
ecm
}
\author{
Matthew W. Pennell
}
\references{
Miller C. R., P. Joyce and L.P. Waits. 2005. A new method for estimating the size of small populations from genetic mark-recapture data. Molecular Ecology 14:1991-2005. Pennell, M. W., C. R. Stansbury, L. P. Waits and C. R. Miller. Capwrie: a R package for estimating population census size from non-invasive genetic sampling
}
\seealso{
\code{\link{fitEcm}}
}

