% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{wombat}
\alias{wombat}
\title{Wombat dataset

This dataset is used to demonstrate the functionality of the package (see supplemental material of Pennell et al. 2013). It can be used as input for \code{\link{fitTirm}} and \code{\link{fitEcm}}.}
\format{data frame}
\description{
Wombat dataset

This dataset is used to demonstrate the functionality of the package (see supplemental material of Pennell et al. 2013). It can be used as input for \code{\link{fitTirm}} and \code{\link{fitEcm}}.
}
\references{
Banks, S. C., S. D. Hoyle, A. Horsup, P. Sunnucks, and A. C. Taylor. 2003. Demographic monitoring of an entire species (the northern hairy-nosed wombat, Lasiorhinus kretii) by genetic analysis of non-invasively collected material. Animal Conservation 6:101-107.
}
\keyword{datasets}

