% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bootstrapCapwire}
\alias{bootstrapCapwire}
\title{Parametric bootstrap ECM and TIRM models}
\usage{
bootstrapCapwire(fit, bootstraps = 1000, CI = c(0.025, 0.975))
}
\arguments{
\item{fit}{a \code{capfit} object inherited from \code{\link{fitEcm}} or \code{\link{fitTirm}}}

\item{bootstraps}{The number of bootstraps to be performed (default is 1000)}

\item{CI}{A vector of quantiles to to generate a confidence interval for the population estimate. The default is \code{c(0.025, 0.975)}, denoting a symmetrical 95 percent confidence interval.}
}
\value{
A list containing the following elements:
\describe{
 \item{ml.pop.size}{The maximum likelihood estimate for the population size obtained by fitting the model}
 \item{conf.int}{The confidence interval for the estimate of the population size}
}
}
\description{
Uses maximum likelihood parameter estimates from \code{fitEcm}, or \code{fitTirm} to perform a parametric bootstrap to get confidence intervals for the estimate of the population size
}
\details{
This function uses the ML estimates obtained from fitting the model to simulate data under the model.

\code{\link{bootstrapCapwire}} inherits an object from \code{\link{fitEcm}} or \code{\link{fitTirm}} such that the model and parameter estimates do not need to be specified.

The ML estimate for the population size will also be returned but this will not be changed by \code{bootstrapCapwire}

Note that if the model is a poor fit to the data, the confidence intervals may not be reliable.

The lower confidence interval is bounded by the number of unique individuals in the sample.
}
\examples{
## Simulate data under Equal Capture Model

data <- simEcm(n=40, s=150)

## Fit Equal Capture Model to Data

res <- fitEcm(data=data, max.pop=200)

## Perform bootstrap to get confidence intervals

ci <- bootstrapCapwire(fit=res, bootstraps=50, CI = c(0.025, 0.975))
ci
}
\author{
Matthew W. Pennell
}
\references{
Miller C. R., P. Joyce and L.P. Waits. 2005. A new method for estimating the size of small populations from genetic mark-recapture data. Molecular Ecology 14:1991-2005.

Pennell, M. W., C. R. Stansbury, L. P. Waits and C. R. Miller. 2013. Capwire: a R package for estimating population census size from non-invasive genetic sampling. Molecular Ecology Resources 13:154-157.
}
\seealso{
\code{\link{fitTirm}}, \code{\link{fitEcm}}
}

