% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawCapRatesExp}
\alias{drawCapRatesExp}
\title{Create a exponential distribution of capture rates for simulation}
\usage{
drawCapRatesExp(r)
}
\arguments{
\item{r}{exponential rate parameter}
}
\value{
a function which takes an argument \code{n} that can be then used to simulate capture rates with \code{\link{simCapture}}
}
\description{
Can be used in conjunction with \code{\link{simCapture}}
}
\examples{
## Specify an exponential distribution
ed <- drawCapRatesExp(0.5)
simCapture(n=20, s=100, ed)
}
\seealso{
\code{\link{simCapture}} \code{\link{rexp}}
}

