% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawCapRatesGeom}
\alias{drawCapRatesGeom}
\title{Create a geometric distribution of capture rates for simulation}
\usage{
drawCapRatesGeom(p)
}
\arguments{
\item{p}{geometric sampling probability}
}
\value{
a function which takes an argument \code{n} that can be then used to simulate capture rates with \code{\link{simCapture}}
}
\description{
Can be used in conjunction with \code{\link{simCapture}}
}
\examples{
## Specify a geometric distribution
md <- drawCapRatesGeom(0.5)
simCapture(n=20, s=100, md)
}
\seealso{
\code{\link{simCapture}} \code{\link{rgeom}}
}

