% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawCapRatesUnif}
\alias{drawCapRatesUnif}
\title{Create a uniform distribution of capture rates for simulation}
\usage{
drawCapRatesUnif(lower, upper)
}
\arguments{
\item{lower}{lower bound of uniform distribution}

\item{upper}{upper bound of uniform distribution}
}
\value{
a function which takes an argument \code{n} that can be then used to simulate capture rates with \code{\link{simCapture}}
}
\description{
Can be used in conjunction with \code{\link{simCapture}}
}
\examples{
## Specify a uniform distribution
ud <- drawCapRatesUnif(0,1)
simCapture(n=20, s=100, ud)
}
\seealso{
\code{\link{simCapture}} \code{\link{runif}}
}

