% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simTirm}
\alias{simTirm}
\title{Simulate data under Two Innate Rates Model (TIRM)}
\usage{
simTirm(na, nb, a, s)
}
\arguments{
\item{na}{Number of individuals in the "easy to capture" class}

\item{nb}{Number of individuals in the "difficult to capture" class}

\item{a}{The ratio of capture rates between the two class (i.e. the alpha parameter in \code{\link{fitTirm}}}

\item{s}{Total number of samples collected}
}
\value{
A two-column matrix with the first column specifiying the capture class (i.e. individuals caught i times) and the second column specifying the number of individuals in each class.

The matrix can be used as the data argument for any of the model-fitting functions implemented in capwire
}
\description{
Simulates capture count data under the assumptions of the Two-Innate Rates Model (TIRM) where the individuals in the population are assumed to come from two classes: easy to capture individuals and difficult to capture individuals.
}
\examples{
## Simualte data under Two-Innate Rates Model

data <- simTirm(na=20, nb=10, a=4, s=100)

## Fit the Two-Innate Rates Model

tirm <- fitTirm(data=data, max.pop=200)
tirm
}
\author{
Matthew W. Pennell
}
\references{
Miller C. R., P. Joyce and L.P. Waits. 2005. A new method for estimating the size of small populations from genetic mark-recapture data. Molecular Ecology 14:1991-2005.

Pennell, M. W., C. R. Stansbury, L. P. Waits and C. R. Miller. 2013. Capwire: a R package for estimating population census size from non-invasive genetic sampling. Molecular Ecology Resources 13:154-157.
}
\seealso{
\code{\link{fitTirm}}
}

